Внесены последние правки: удалён лишний 'run', убраны костыли с 'fontSize', экраны добавлены в пакеты с вью моделями
