+++ app/src/main/java/com/dz/bmstu_trade/ui/main/connect/device_code/AddDeviceViewModel.kt
@@ -1,4 +1,4 @@
-package com.dz.bmstu_trade.ui.main.connect.deviceCodeVM
+package com.dz.bmstu_trade.ui.main.connect.device_code
 
 
 import androidx.annotation.StringRes
