+++ app/src/main/java/com/dz/bmstu_trade/ui/main/connect/device_code/DeviceManualConnectScreen.kt
@@ -1,4 +1,4 @@
-package com.dz.bmstu_trade.ui.main.connect
+package com.dz.bmstu_trade.ui.main.connect.device_code
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -22,27 +22,27 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.dz.bmstu_trade.ui.main.connect.deviceNamingVM.NameDeviceViewModel
+import androidx.navigation.NavHostController
 import com.dz.bmstu_trade.R
-import com.dz.bmstu_trade.ui.main.connect.deviceNamingVM.NameFieldState
+
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun DeviceNamingScreen(
-    viewModel: NameDeviceViewModel = remember { NameDeviceViewModel() }
+fun DeviceManualConnectScreen(
+    navController: NavHostController? = null,
+    viewModel: AddDeviceViewModel = remember { AddDeviceViewModel() }
 ) {
-    val nameFieldState by viewModel.name.collectAsState()
+    val codeFieldState by viewModel.code.collectAsState()
 
     Column (
         modifier = Modifier
             .fillMaxSize()
     ) {
         TopAppBar(
-            title = { Text(text = stringResource(R.string.naming_top_bar_title)) },
+            title = { Text(text = stringResource(R.string.manual_connection_top_bar_title)) },
             colors = TopAppBarDefaults.topAppBarColors(
                 containerColor = MaterialTheme.colorScheme.secondaryContainer,
                 titleContentColor = MaterialTheme.colorScheme.onSurface,
@@ -51,7 +51,7 @@ fun DeviceNamingScreen(
                 IconButton(onClick = { /*TODO*/ }) {
                     Icon(
                         imageVector = Icons.Default.ArrowBack,
-                        contentDescription = stringResource(R.string.arrow_back_icon_description),
+                        contentDescription = stringResource(R.string.arrow_back_icon_description)
                     )
                 }
             }
@@ -67,43 +67,49 @@ fun DeviceNamingScreen(
             contentAlignment = Alignment.Center
         ) {
             Column {
-                Text(text = stringResource(R.string.naming_rules_label))
+                Text(text = stringResource(R.string.code_location_label))
                 OutlinedTextField(
                     modifier = Modifier
                         .fillMaxWidth()
                         .padding(top = dimensionResource(R.dimen.space_between_inputs_texts_buttons)),
-                    value = nameFieldState.value,
-                    onValueChange = { viewModel.onNameUpdated(it) },
-                    isError = nameFieldState.error != null,
-                    label = { Text(text = stringResource(R.string.device_name_label)) }
+                    value = codeFieldState.value,
+                    onValueChange = { viewModel.onCodeUpdated(it) },
+                    isError = codeFieldState.error != null
+                            && codeFieldState.error != TextFieldState.Error.TOO_SHORT,
+                    label = { Text(text = stringResource(R.string.input_device_code_label))}
                 )
-
-                if (nameFieldState.error != null) {
+                if (codeFieldState.error != null) {
                     Text(
-                        text = stringResource(nameFieldState.error!!.messageResId),
+                        text = stringResource(codeFieldState.error!!.messageResId),
                         modifier = Modifier
                             .padding(top = dimensionResource(R.dimen.space_between_inputs_texts_buttons)),
-                        color = MaterialTheme.colorScheme.error,
+                        color = when (codeFieldState.error) {
+                            TextFieldState.Error.TOO_LARGE -> {
+                                MaterialTheme.colorScheme.error
+                            }
+                            TextFieldState.Error.TOO_SHORT -> {
+                                MaterialTheme.colorScheme.inverseSurface
+                            }
+                            else -> MaterialTheme.colorScheme.onSurface
+                        },
                     )
                 }
-
                 Button(
-                    enabled = nameFieldState.error == null,
+                    enabled =  codeFieldState.error == null,
                     modifier = Modifier
                         .fillMaxWidth()
                         .padding(top = dimensionResource(R.dimen.space_between_inputs_texts_buttons)),
                     onClick = { /* TODO */ }
                 ) {
-                    Text(text = stringResource(id = R.string.apply_label))
+                    Text(text = stringResource(id = R.string.connect_by_code_button_title))
                 }
             }
         }
     }
 }
 
-
 @Preview(showBackground = true)
 @Composable
-fun DeviceNamingPreview() {
-    DeviceNamingScreen()
+fun ManualConnectionPreview() {
+    DeviceManualConnectScreen()
 }
